Created a util file for register and some formatting changes
