+++ cmd/register/register_test.go
@@ -9,8 +9,8 @@ import (
 
 func TestRegisterCommand(t *testing.T) {
 	registerCommand := RegisterCommand()
-	assert.Equal(t, registerCommand.Use , "register")
-	assert.Equal(t, registerCommand.Short , "Registers tasks/workflows/launchplans from list of generated serialized files.")
+	assert.Equal(t, registerCommand.Use, "register")
+	assert.Equal(t, registerCommand.Short, "Registers tasks/workflows/launchplans from list of generated serialized files.")
 	fmt.Println(registerCommand.Commands())
 	assert.Equal(t, len(registerCommand.Commands()), 1)
 	cmdNouns := registerCommand.Commands()
